requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

annotated-types==0.5.0
anyio==3.7.1
arcgis==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.3
attrs==23.1.0
azure-functions==1.15.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
fastjsonschema==2.18.0
fonttools==4.41.1
fqdn==1.5.1
geomet==1.0.0
gssapi==1.8.2
idna==3.4
ipykernel==6.25.0
ipython==8.14.0
ipywidgets==8.0.7
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keyring==21.7.0
kiwisolver==1.4.4
lerc==4.0.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.1
nest-asyncio==1.5.6
notebook_shim==0.2.3
numpy==1.25.1
oauthlib==3.2.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.15.1
pyparsing==3.0.9
pypiwin32==223
pyshp==2.3.1
pyspnego==0.9.1
python-certifi-win32==1.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.0
requests==2.31.0
requests-gssapi==1.2.3
requests-kerberos==0.14.0
requests-negotiate-sspi==0.5.2
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
Send2Trash==1.8.2
setuptools-scm==7.1.0
six==1.16.0
smartsheet-python-sdk==3.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
urllib3==1.25.11
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.8
winkerberos==0.9.1
wrapt==1.15.0
